# trade-capture-service/src/main/java/com/pbsynth/tradecapture/dto/CaptureTradeAcceptedResponse.java
package com.pbsynth.tradecapture.dto;

import java.time.Instant;

public record CaptureTradeAcceptedResponse(
        String ingestionId,
        String status,
        Instant acceptedAt,
        String lifecycleInstructionType
) {}